New examples
    Changes in the stimuli library
    FIx Atrial mesh domain function
    Ignoring private files

+++ test_headers/test_headers.sh
@@ -3,14 +3,18 @@ while IFS='' read -r line || [[ -n "$line" ]]; do
 
     error_file=$(basename "${line}")
 
-    if cc  -I/opt/cuda/include/ -Werror "${line}" -o tmp.gch > "${error_file}".txt
-    then
-        echo Header "$line" compiled sucessfuly!
-        rm "${error_file}".txt
-    else
-        echo Failed to compile header "$line"
-    fi
+	if test -f "$line"; then
 
-    rm tmp.gch
+		if cc  -I/opt/cuda/include/ -Werror "${line}" -o tmp.gch > "${error_file}".txt
+		then
+			echo Header "$line" compiled sucessfuly!
+			rm "${error_file}".txt
+		else
+			echo Failed to compile header "$line"
+		fi
+
+		rm tmp.gch
+
+	fi
 
 done < "$1"
